refactor: migrate app routing to RouterProvider data router
